Material mapping in velocity field and implementation in stack redeposition

+++ include/Models/SF6O2Etching.hpp
@@ -65,8 +65,7 @@ public:
         break;
       }
 
-      if (psMaterialMap::mapToMaterial(static_cast<int>(materialIds[i])) ==
-          psMaterial::Si) {
+      if (psMaterialMap::mapToMaterial(materialIds[i]) == psMaterial::Si) {
 
         etchRate[i] =
             -(1 / rho_Si) *
